third_party/blink/web_tests/accessibility/continuation2.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<foo>
  <bar>
    <div id="before">Before</div>
  </bar><div id="after">After</div>
</foo>

<script>
test(function(t) {
    var axBefore = accessibilityController.accessibleElementById("before");
    assert_equals(axBefore.childAtIndex(0).name, "Before");
    var axAfter = accessibilityController.accessibleElementById("after");
    assert_equals(axAfter.childAtIndex(0).name, "After");
}, "Ensure that continuations are included in the accessibility tree.");
</script>
